project/api/requirements.txt
requests
PyMySQL
PyJWT
django
djangorestframework
django-cors-headers
Pillow